requirements.txt
langchain==0.0.292
openai==0.28.0
datasets==2.10.1
pinecone-client==2.2.4
tiktoken==0.5.1
python-dotenv==1.0.1